Refactor(EventObserver): Update `success` method signature to use `EventDataStatus`
The `success` method within `ObserveWrapper` has been updated.
Previously, it accepted a `withData: Boolean` parameter. This has been replaced with `dataStatus: EventDataStatus`.

This change allows for more granular control over when the success event is triggered based on the presence or absence of data, using the `EventDataStatus` enum (which can be `WithData`, `WithoutData`, or `DoesNotMatter`).

The default value for `dataStatus` is `EventDataStatus.DoesNotMatter`.